🧬 Migrate the logic about generating Python code to another single module *pymock_api.server.application.code_generator*. (not finish yet)
