+++ pymock_api/server/application/__init__.py
@@ -13,6 +13,11 @@ from typing import Any, Dict, List, Optional, Union, cast
 from ..._utils import import_web_lib
 from ...exceptions import FileFormatNotSupport
 from ...model import APIParameter, HTTPRequest, HTTPResponse, MockAPI, MockAPIs
+from .code_generator import (
+    BaseWebServerCodeGenerator,
+    FastAPICodeGenerator,
+    FlaskCodeGenerator,
+)
 
 
 class BaseAppServer(metaclass=ABCMeta):
@@ -29,12 +34,21 @@ class BaseAppServer(metaclass=ABCMeta):
         self._mock_api_details: Dict[str, Dict[str, MockAPI]] = {}
         self._api_functions: Dict[str, str] = {}
 
+        self._code_generator = self.init_code_generator()
+
     @property
     def web_application(self) -> Any:
         """:obj:`Any`: Property with only getter for the instance of web application, e.g., *Flask*, *FastAPI*, etc."""
         if not self._web_application:
             self._web_application = self.setup()
         return self._web_application
+
+    @property
+    def mock_api_details(self) -> Dict[str, Dict[str, MockAPI]]:
+        if not self._mock_api_details:
+            self._mock_api_details = getattr(self._code_generator, "_mock_api_details")
+            return self._mock_api_details
+        return self._mock_api_details
 
     @abstractmethod
     def setup(self) -> Any:
@@ -46,6 +60,10 @@ class BaseAppServer(metaclass=ABCMeta):
         """
         pass
 
+    @abstractmethod
+    def init_code_generator(self) -> BaseWebServerCodeGenerator:
+        pass
+
     def create_api(self, mocked_apis: MockAPIs) -> None:
         """
         [Entry point for generating Python code]
@@ -53,8 +71,8 @@ class BaseAppServer(metaclass=ABCMeta):
         aggregated_mocked_apis = self._get_all_api_details(mocked_apis)
         for api_name, api_config in aggregated_mocked_apis.items():
             if api_name and api_config:
-                annotate_function_pycode = self._annotate_function(api_name, api_config)  # type: ignore[arg-type]
-                add_api_pycode = self._add_api(
+                annotate_function_pycode = self._code_generator.annotate_function(api_name, api_config)
+                add_api_pycode = self._code_generator.add_api(
                     api_name, api_config, base_url=mocked_apis.base.url if mocked_apis.base else None
                 )
                 print(f"[DEBUG in src] annotate_function_pycode: {annotate_function_pycode}")
@@ -71,49 +89,49 @@ class BaseAppServer(metaclass=ABCMeta):
         """
         pass
 
-    def _annotate_function(self, api_name: str, api_config: List[MockAPI]) -> str:
-        """
-        [Generating code]
-        """
-        initial_global_server = f"""global SERVER\nSERVER = self\n"""
-        define_function_for_api = self._define_api_function_pycode(api_name, api_config)
-        return initial_global_server + define_function_for_api
-
-    def _define_api_function_pycode(self, api_name: str, api_config: List[MockAPI]) -> str:
-        """
-        [Generating code]
-        """
-        api_function_name = "_".join(api_name.split("/")[1:]).replace("-", "_")
-        return f"""def {api_function_name}() -> Union[str, dict]:
-            {self._run_request_process_pycode()}
-            {self._handle_request_process_result_pycode()}
-            {self._generate_response_pycode()}
-        """
-
-    def _run_request_process_pycode(self, **kwargs) -> str:
-        """
-        [Generating code]
-        """
-        return """
-        process_result = SERVER._request_process()
-        """
-
-    def _handle_request_process_result_pycode(self, **kwargs) -> str:
-        """
-        [Generating code]
-        """
-        return """
-        if process_result.status_code != 200:
-            return process_result
-        """
-
-    def _generate_response_pycode(self) -> str:
-        """
-        [Generating code]
-        """
-        return f"""
-        return SERVER._response_process()
-        """
+    # def _annotate_function(self, api_name: str, api_config: List[MockAPI]) -> str:
+    #     """
+    #     [Generating code]
+    #     """
+    #     initial_global_server = f"""global SERVER\nSERVER = self\n"""
+    #     define_function_for_api = self._define_api_function_pycode(api_name, api_config)
+    #     return initial_global_server + define_function_for_api
+    #
+    # def _define_api_function_pycode(self, api_name: str, api_config: List[MockAPI]) -> str:
+    #     """
+    #     [Generating code]
+    #     """
+    #     api_function_name = "_".join(api_name.split("/")[1:]).replace("-", "_")
+    #     return f"""def {api_function_name}() -> Union[str, dict]:
+    #         {self._run_request_process_pycode()}
+    #         {self._handle_request_process_result_pycode()}
+    #         {self._generate_response_pycode()}
+    #     """
+    #
+    # def _run_request_process_pycode(self, **kwargs) -> str:
+    #     """
+    #     [Generating code]
+    #     """
+    #     return """
+    #     process_result = SERVER._request_process()
+    #     """
+    #
+    # def _handle_request_process_result_pycode(self, **kwargs) -> str:
+    #     """
+    #     [Generating code]
+    #     """
+    #     return """
+    #     if process_result.status_code != 200:
+    #         return process_result
+    #     """
+    #
+    # def _generate_response_pycode(self) -> str:
+    #     """
+    #     [Generating code]
+    #     """
+    #     return f"""
+    #     return SERVER._response_process()
+    #     """
 
     def _ensure_http(self, api_config: MockAPI, http_attr: str) -> Union[HTTPRequest, HTTPResponse]:
         """
@@ -124,43 +142,43 @@ class BaseAppServer(metaclass=ABCMeta):
         ), "The configuration *HTTP* value should not be empty."
         return getattr(api_config.http, http_attr)
 
-    @abstractmethod
-    def _add_api(self, api_name: str, api_config: Union[MockAPI, List[MockAPI]], base_url: Optional[str] = None) -> str:
-        """
-        [Generating code] but doing data processing first
-        """
-        if isinstance(api_config, list):
-            url = api_name
-        elif isinstance(api_config, MockAPI):
-            url = api_config.url  # type: ignore[assignment]
-        else:
-            raise TypeError
-        self._record_api_params_info(url=self.url_path(url=url, base_url=base_url), api_config=api_config)
-        return ""
-
-    def url_path(self, url: Optional[str], base_url: Optional[str] = None) -> str:
-        """
-        [Data processing]
-        """
-        assert url
-        return f"{base_url}{url}" if base_url else f"{url}"
-
-    def _record_api_params_info(self, url: str, api_config: Union[MockAPI, List[MockAPI]]) -> None:
-        """
-        [Generating code] but doing data processing first
-        This processing for the outer function which would be called by the generating code
-        """
-        if isinstance(api_config, list):
-            for ac in api_config:
-                url_details = self._mock_api_details.get(url, {})
-                url_details[ac.http.request.method] = ac  # type: ignore[union-attr]
-                self._mock_api_details[url] = url_details
-        elif isinstance(api_config, MockAPI):
-            url_details = self._mock_api_details.get(url, {})
-            url_details[api_config.http.request.method] = api_config  # type: ignore[union-attr]
-            self._mock_api_details[url] = url_details
-        else:
-            raise TypeError("")
+    # @abstractmethod
+    # def _add_api(self, api_name: str, api_config: Union[MockAPI, List[MockAPI]], base_url: Optional[str] = None) -> str:
+    #     """
+    #     [Generating code] but doing data processing first
+    #     """
+    #     if isinstance(api_config, list):
+    #         url = api_name
+    #     elif isinstance(api_config, MockAPI):
+    #         url = api_config.url  # type: ignore[assignment]
+    #     else:
+    #         raise TypeError
+    #     self._record_api_params_info(url=self.url_path(url=url, base_url=base_url), api_config=api_config)
+    #     return ""
+    #
+    # def url_path(self, url: Optional[str], base_url: Optional[str] = None) -> str:
+    #     """
+    #     [Data processing]
+    #     """
+    #     assert url
+    #     return f"{base_url}{url}" if base_url else f"{url}"
+    #
+    # def _record_api_params_info(self, url: str, api_config: Union[MockAPI, List[MockAPI]]) -> None:
+    #     """
+    #     [Generating code] but doing data processing first
+    #     This processing for the outer function which would be called by the generating code
+    #     """
+    #     if isinstance(api_config, list):
+    #         for ac in api_config:
+    #             url_details = self._mock_api_details.get(url, {})
+    #             url_details[ac.http.request.method] = ac  # type: ignore[union-attr]
+    #             self._mock_api_details[url] = url_details
+    #     elif isinstance(api_config, MockAPI):
+    #         url_details = self._mock_api_details.get(url, {})
+    #         url_details[api_config.http.request.method] = api_config  # type: ignore[union-attr]
+    #         self._mock_api_details[url] = url_details
+    #     else:
+    #         raise TypeError("")
 
     def _request_process(self, **kwargs) -> "flask.Response":  # type: ignore
         """
@@ -169,7 +187,7 @@ class BaseAppServer(metaclass=ABCMeta):
         request = self._get_current_request(**kwargs)
         req_params = self._get_current_api_parameters(**kwargs)
 
-        api_params_info: List[APIParameter] = self._mock_api_details[self._get_current_api_path(request)][self._get_current_request_http_method(request)].http.request.parameters  # type: ignore[union-attr]
+        api_params_info: List[APIParameter] = self.mock_api_details[self._get_current_api_path(request)][self._get_current_request_http_method(request)].http.request.parameters  # type: ignore[union-attr]
         for param_info in api_params_info:
             # Check the required parameter
             one_req_param_value = req_params.get(param_info.name, None)
@@ -199,7 +217,7 @@ class BaseAppServer(metaclass=ABCMeta):
         [Response processing for generating code]
         """
         request = self._get_current_request(**kwargs)
-        api_params_info: MockAPI = self._mock_api_details[self._get_current_api_path(request)][
+        api_params_info: MockAPI = self.mock_api_details[self._get_current_api_path(request)][
             self._get_current_request_http_method(request)
         ]
         response_value = self._ensure_http(api_params_info, "response").value  # type: ignore[union-attr]
@@ -240,12 +258,12 @@ class BaseAppServer(metaclass=ABCMeta):
         """
         pass
 
-    @abstractmethod
-    def _api_controller_name(self, api_name: str) -> str:
-        """
-        [Generating code] but doing data processing about naming function
-        """
-        pass
+    # @abstractmethod
+    # def _api_controller_name(self, api_name: str) -> str:
+    #     """
+    #     [Generating code] but doing data processing about naming function
+    #     """
+    #     pass
 
 
 class FlaskServer(BaseAppServer):
@@ -253,6 +271,9 @@ class FlaskServer(BaseAppServer):
 
     def setup(self) -> "flask.Flask":  # type: ignore
         return import_web_lib.flask().Flask(__name__)
+
+    def init_code_generator(self) -> FlaskCodeGenerator:
+        return FlaskCodeGenerator()
 
     def _get_all_api_details(self, mocked_apis: MockAPIs) -> Dict[str, List[MockAPI]]:  # type: ignore[override]
         return mocked_apis.group_by_url()
@@ -276,19 +297,19 @@ class FlaskServer(BaseAppServer):
     def _generate_http_response(self, body: str, status_code: int) -> "flask.Response":  # type: ignore
         return import_web_lib.flask().Response(body, status=status_code)
 
-    def _add_api(self, api_name: str, api_config: List[MockAPI], base_url: Optional[str] = None) -> str:  # type: ignore[override]
-        super()._add_api(api_name=api_name, api_config=api_config, base_url=base_url)
-        # TODO: Should align the data structure and remove this checking
-        if not isinstance(api_config, list):
-            raise TypeError("")
-        acceptance_method = [cast(HTTPRequest, self._ensure_http(ac, "request")).method for ac in api_config]
-        return f"""self.web_application.route(
-            "{self.url_path(api_name, base_url)}", methods={acceptance_method}
-            )({self._api_controller_name(api_name)})
-        """
-
-    def _api_controller_name(self, api_name: str) -> str:
-        return "_".join(api_name.split("/")[1:]).replace("-", "_")
+    # def _add_api(self, api_name: str, api_config: List[MockAPI], base_url: Optional[str] = None) -> str:  # type: ignore[override]
+    #     super()._add_api(api_name=api_name, api_config=api_config, base_url=base_url)
+    #     # TODO: Should align the data structure and remove this checking
+    #     if not isinstance(api_config, list):
+    #         raise TypeError("")
+    #     acceptance_method = [cast(HTTPRequest, self._ensure_http(ac, "request")).method for ac in api_config]
+    #     return f"""self.web_application.route(
+    #         "{self.url_path(api_name, base_url)}", methods={acceptance_method}
+    #         )({self._api_controller_name(api_name)})
+    #     """
+    #
+    # def _api_controller_name(self, api_name: str) -> str:
+    #     return "_".join(api_name.split("/")[1:]).replace("-", "_")
 
 
 class FastAPIServer(BaseAppServer):
@@ -301,118 +322,121 @@ class FastAPIServer(BaseAppServer):
     def setup(self) -> "fastapi.FastAPI":  # type: ignore
         return import_web_lib.fastapi().FastAPI()
 
+    def init_code_generator(self) -> FastAPICodeGenerator:
+        return FastAPICodeGenerator()
+
     def _get_all_api_details(self, mocked_apis: MockAPIs) -> Dict[str, Optional[MockAPI]]:  # type: ignore[override]
         return mocked_apis.apis
 
-    def _annotate_function(self, api_name: str, api_config: MockAPI) -> str:  # type: ignore[override]
-        import_fastapi = "from fastapi import Request as FastAPIRequest\n"
-        initial_global_server = f"""global SERVER\nSERVER = self\n"""
-        define_params_model = self._annotate_api_parameters_model_pycode(api_name, api_config)
-        define_function_for_api = self._define_api_function_pycode(api_name, api_config)
-        return import_fastapi + initial_global_server + define_params_model + define_function_for_api
-
-    def _define_api_function_pycode(self, api_name: str, api_config: MockAPI) -> str:  # type: ignore[override]
-        # The code implementation is different if the HTTP method is *GET* or not
-        if api_config.http.request.method.upper() != "GET":  # type: ignore[union-attr]
-            api_func_signature = ""
-            # Process the function signature if API has parameter settings
-            if api_config.http.request.parameters:  # type: ignore[union-attr]
-                parameter_class = self._api_name_as_camel_case(api_name)
-                api_func_signature = f"model: {parameter_class}, " if self._api_has_params else ""
-            # Combine all the string value as a valid Python code
-            return f"""def {api_name}({api_func_signature}request: FastAPIRequest):
-                {self._run_request_process_pycode()}
-                {self._handle_request_process_result_pycode()}
-                {self._generate_response_pycode()}
-            """
-        else:
-            function_args = ""
-            assign_value_to_model = ""
-            instantiate_model = ""
-            # Process the function signature if API has parameter settings
-            for param in api_config.http.request.parameters:  # type: ignore[union-attr]
-                if param.default is not None:
-                    if param.value_type == "str":
-                        default_value = f"'{param.default}'"
-                    else:
-                        default_value = f"{param.default}"
-                    function_args += f", {param.name}: {param.value_type} = {default_value}"
-                else:
-                    function_args += f", {param.name}: {param.value_type} = None"
-                assign_value_to_model += f"""
-        setattr(model, '{param.name}', {param.name})
-                """
-            # Instantiate and assign data into objects as data model
-            if api_config.http.request.parameters:  # type: ignore[union-attr]
-                parameter_class = self._api_name_as_camel_case(api_name)
-                instantiate_model = f"""
-        class {parameter_class}: pass
-        model = {parameter_class}()
-                """
-            # Combine all the string value as a valid Python code
-            return f"""def {api_name}(request: FastAPIRequest{function_args}):
-                {instantiate_model}
-                {assign_value_to_model}
-                {self._run_request_process_pycode()}
-                {self._handle_request_process_result_pycode()}
-                {self._generate_response_pycode()}
-            """
-
-    def _api_name_as_camel_case(self, api_name: str) -> str:
-        # api_function_name = "_".join(api_name.split("/")[1:]).replace("-", "_")
-        # camel_case_api_name = "".join(map(lambda n: f"{n[0].upper()}{n[1:]}", api_function_name.split("_")))
-        camel_case_api_name = "".join(map(lambda n: f"{n[0].upper()}{n[1:]}", api_name.split("_")))
-        return f"{camel_case_api_name}Parameter"
-
-    def _annotate_api_parameters_model_pycode(self, api_name: str, api_config: MockAPI) -> str:
-        define_parameters_model = ""
-        if api_config.http.request.parameters:  # type: ignore[union-attr]
-            # The code implementation is different if the HTTP method is *GET* or not
-            if api_config.http.request.method.upper() == "GET":  # type: ignore[union-attr]
-                # API with HTTP method *GET* doesn't need to use 'pydantic.BaseModel' object to process API parameters
-                self._api_has_params = True
-                return define_parameters_model
-            else:
-                # API without HTTP method *GET* needs to use 'pydantic.BaseModel' object to process API parameters
-                self._api_has_params = True
-                # Handle the class annotation
-                parameter_class = self._api_name_as_camel_case(api_name)
-                define_parameters_model = f"""from pydantic import BaseModel\nclass {parameter_class}(BaseModel):\n"""
-            # Handle the class's attributes
-            for prop in api_config.http.request.parameters:  # type: ignore[union-attr]
-                if prop.default is not None:
-                    define_parameters_model += f"    {prop.name}: {prop.value_type} = '{prop.default}'\n"
-                else:
-                    if prop.required:
-                        define_parameters_model += f"    {prop.name}: {prop.value_type}\n"
-                    else:
-                        define_parameters_model += f"    {prop.name}: {prop.value_type} = None\n"
-            return define_parameters_model
-        else:
-            self._api_has_params = False
-            return define_parameters_model
-
-    def _run_request_process_pycode(self, **kwargs) -> str:
-        return (
-            """
-        process_result = SERVER._request_process(model=model, request=request)
-        """
-            if self._api_has_params
-            else """
-        process_result = SERVER._request_process(request=request)
-        """
-        )
-
-    def _generate_response_pycode(self) -> str:
-        return f"""
-        return SERVER._response_process(request=request)
-        """
+    # def _annotate_function(self, api_name: str, api_config: MockAPI) -> str:  # type: ignore[override]
+    #     import_fastapi = "from fastapi import Request as FastAPIRequest\n"
+    #     initial_global_server = f"""global SERVER\nSERVER = self\n"""
+    #     define_params_model = self._annotate_api_parameters_model_pycode(api_name, api_config)
+    #     define_function_for_api = self._define_api_function_pycode(api_name, api_config)
+    #     return import_fastapi + initial_global_server + define_params_model + define_function_for_api
+    #
+    # def _define_api_function_pycode(self, api_name: str, api_config: MockAPI) -> str:  # type: ignore[override]
+    #     # The code implementation is different if the HTTP method is *GET* or not
+    #     if api_config.http.request.method.upper() != "GET":  # type: ignore[union-attr]
+    #         api_func_signature = ""
+    #         # Process the function signature if API has parameter settings
+    #         if api_config.http.request.parameters:  # type: ignore[union-attr]
+    #             parameter_class = self._api_name_as_camel_case(api_name)
+    #             api_func_signature = f"model: {parameter_class}, " if self._api_has_params else ""
+    #         # Combine all the string value as a valid Python code
+    #         return f"""def {api_name}({api_func_signature}request: FastAPIRequest):
+    #             {self._run_request_process_pycode()}
+    #             {self._handle_request_process_result_pycode()}
+    #             {self._generate_response_pycode()}
+    #         """
+    #     else:
+    #         function_args = ""
+    #         assign_value_to_model = ""
+    #         instantiate_model = ""
+    #         # Process the function signature if API has parameter settings
+    #         for param in api_config.http.request.parameters:  # type: ignore[union-attr]
+    #             if param.default is not None:
+    #                 if param.value_type == "str":
+    #                     default_value = f"'{param.default}'"
+    #                 else:
+    #                     default_value = f"{param.default}"
+    #                 function_args += f", {param.name}: {param.value_type} = {default_value}"
+    #             else:
+    #                 function_args += f", {param.name}: {param.value_type} = None"
+    #             assign_value_to_model += f"""
+    #     setattr(model, '{param.name}', {param.name})
+    #             """
+    #         # Instantiate and assign data into objects as data model
+    #         if api_config.http.request.parameters:  # type: ignore[union-attr]
+    #             parameter_class = self._api_name_as_camel_case(api_name)
+    #             instantiate_model = f"""
+    #     class {parameter_class}: pass
+    #     model = {parameter_class}()
+    #             """
+    #         # Combine all the string value as a valid Python code
+    #         return f"""def {api_name}(request: FastAPIRequest{function_args}):
+    #             {instantiate_model}
+    #             {assign_value_to_model}
+    #             {self._run_request_process_pycode()}
+    #             {self._handle_request_process_result_pycode()}
+    #             {self._generate_response_pycode()}
+    #         """
+    #
+    # def _api_name_as_camel_case(self, api_name: str) -> str:
+    #     # api_function_name = "_".join(api_name.split("/")[1:]).replace("-", "_")
+    #     # camel_case_api_name = "".join(map(lambda n: f"{n[0].upper()}{n[1:]}", api_function_name.split("_")))
+    #     camel_case_api_name = "".join(map(lambda n: f"{n[0].upper()}{n[1:]}", api_name.split("_")))
+    #     return f"{camel_case_api_name}Parameter"
+    #
+    # def _annotate_api_parameters_model_pycode(self, api_name: str, api_config: MockAPI) -> str:
+    #     define_parameters_model = ""
+    #     if api_config.http.request.parameters:  # type: ignore[union-attr]
+    #         # The code implementation is different if the HTTP method is *GET* or not
+    #         if api_config.http.request.method.upper() == "GET":  # type: ignore[union-attr]
+    #             # API with HTTP method *GET* doesn't need to use 'pydantic.BaseModel' object to process API parameters
+    #             self._api_has_params = True
+    #             return define_parameters_model
+    #         else:
+    #             # API without HTTP method *GET* needs to use 'pydantic.BaseModel' object to process API parameters
+    #             self._api_has_params = True
+    #             # Handle the class annotation
+    #             parameter_class = self._api_name_as_camel_case(api_name)
+    #             define_parameters_model = f"""from pydantic import BaseModel\nclass {parameter_class}(BaseModel):\n"""
+    #         # Handle the class's attributes
+    #         for prop in api_config.http.request.parameters:  # type: ignore[union-attr]
+    #             if prop.default is not None:
+    #                 define_parameters_model += f"    {prop.name}: {prop.value_type} = '{prop.default}'\n"
+    #             else:
+    #                 if prop.required:
+    #                     define_parameters_model += f"    {prop.name}: {prop.value_type}\n"
+    #                 else:
+    #                     define_parameters_model += f"    {prop.name}: {prop.value_type} = None\n"
+    #         return define_parameters_model
+    #     else:
+    #         self._api_has_params = False
+    #         return define_parameters_model
+    #
+    # def _run_request_process_pycode(self, **kwargs) -> str:
+    #     return (
+    #         """
+    #     process_result = SERVER._request_process(model=model, request=request)
+    #     """
+    #         if self._api_has_params
+    #         else """
+    #     process_result = SERVER._request_process(request=request)
+    #     """
+    #     )
+    #
+    # def _generate_response_pycode(self) -> str:
+    #     return f"""
+    #     return SERVER._response_process(request=request)
+    #     """
 
     def _get_current_request(self, **kwargs) -> "fastapi.Request":  # type: ignore[name-defined]
         return kwargs.get("request")
 
     def _get_current_api_parameters(self, **kwargs) -> dict:
-        api_params_info: List[APIParameter] = self._mock_api_details[self._get_current_api_path(kwargs["request"])][self._get_current_request_http_method(kwargs["request"])].http.request.parameters  # type: ignore[union-attr]
+        api_params_info: List[APIParameter] = self.mock_api_details[self._get_current_api_path(kwargs["request"])][self._get_current_request_http_method(kwargs["request"])].http.request.parameters  # type: ignore[union-attr]
         api_param_names = list(map(lambda e: e.name, api_params_info))
         api_param = {}
         if "model" in kwargs.keys():
@@ -430,19 +454,19 @@ class FastAPIServer(BaseAppServer):
     def _generate_http_response(self, body: str, status_code: int) -> "fastapi.Response":  # type: ignore
         return import_web_lib.fastapi().Response(body, status_code=status_code)
 
-    def _add_api(self, api_name: str, api_config: MockAPI, base_url: Optional[str] = None) -> str:  # type: ignore[override]
-        super()._add_api(api_name=api_name, api_config=api_config, base_url=base_url)
-        # TODO: Should align the data structure and remove this checking
-        if not isinstance(api_config, MockAPI):
-            raise TypeError("")
-        http_method = api_config.http.request.method.lower()  # type: ignore[union-attr]
-        url_path = self.url_path(api_config.url, base_url)
-        return f"""self.web_application.{http_method}(
-            path="{url_path}")({self._api_controller_name(api_name)})
-        """
-
-    def _api_controller_name(self, api_name: str) -> str:
-        return api_name
+    # def _add_api(self, api_name: str, api_config: MockAPI, base_url: Optional[str] = None) -> str:  # type: ignore[override]
+    #     super()._add_api(api_name=api_name, api_config=api_config, base_url=base_url)
+    #     # TODO: Should align the data structure and remove this checking
+    #     if not isinstance(api_config, MockAPI):
+    #         raise TypeError("")
+    #     http_method = api_config.http.request.method.lower()  # type: ignore[union-attr]
+    #     url_path = self.url_path(api_config.url, base_url)
+    #     return f"""self.web_application.{http_method}(
+    #         path="{url_path}")({self._api_controller_name(api_name)})
+    #     """
+    #
+    # def _api_controller_name(self, api_name: str) -> str:
+    #     return api_name
 
 
 class _HTTPResponse:
